# openapi/reviews.json
{
  "components": {
    "schemas": {
      "CreateReviewDto": {
        "properties": {
          "body": {
            "example": "Comfortable material and true to size.",
            "type": "string"
          },
          "rating": {
            "description": "Star rating 1–5 integer.",
            "example": 5,
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          },
          "title": {
            "example": "Great fit and feel",
            "maxLength": 120,
            "type": "string"
          }
        },
        "required": [
          "rating",
          "title",
          "body"
        ],
        "type": "object"
      },
      "EligibilityResponseDto": {
        "properties": {
          "canReview": {
            "description": "True if customer has confirmed purchases and no existing review for this product.",
            "example": true,
            "type": "boolean"
          },
          "existing": {
            "description": "Alias for existingReviewId.",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "nullable": true,
            "type": "object"
          },
          "existingReviewId": {
            "description": "Existing review ID if customer already reviewed this product.",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "nullable": true,
            "type": "object"
          },
          "reason": {
            "example": "Eligible to review",
            "type": "string"
          }
        },
        "required": [
          "canReview",
          "reason"
        ],
        "type": "object"
      },
      "ModerateDto": {
        "properties": {
          "moderationNote": {
            "description": "Optional internal moderation note.",
            "example": "Content adheres to community guidelines",
            "type": "string"
          }
        },
        "type": "object"
      },
      "PaginatedReviewsResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicReviewDto"
            },
            "type": "array"
          },
          "limit": {
            "example": 10,
            "type": "number"
          },
          "page": {
            "example": 1,
            "type": "number"
          },
          "total": {
            "example": 42,
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "type": "object"
      },
      "PublicReviewDto": {
        "properties": {
          "authorName": {
            "example": "Alex M.",
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "rating": {
            "example": 5,
            "type": "number"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "rating",
          "title",
          "body",
          "authorName",
          "createdAt"
        ],
        "type": "object"
      },
      "ReviewResponseDto": {
        "properties": {
          "authorName": {
            "example": "Alex Doe",
            "type": "string"
          },
          "body": {
            "example": "Comfortable material and true to size.",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customerId": {
            "example": "c3d4e5f6-a7b8-9012-cdef-123456789012",
            "type": "string"
          },
          "id": {
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "type": "string"
          },
          "moderatedAt": {
            "nullable": true,
            "type": "object"
          },
          "moderationNote": {
            "nullable": true,
            "type": "object"
          },
          "orderId": {
            "example": "d4e5f6a7-b8c9-0123-def1-234567890123",
            "type": "string"
          },
          "productId": {
            "example": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
            "type": "string"
          },
          "rating": {
            "example": 5,
            "type": "number"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "example": "approved",
            "type": "string"
          },
          "title": {
            "example": "Great fit and feel",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "example": 1,
            "type": "number"
          }
        },
        "required": [
          "id",
          "productId",
          "customerId",
          "orderId",
          "rating",
          "title",
          "body",
          "authorName",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "UpdateReviewDto": {
        "properties": {
          "body": {
            "example": "Updated review content.",
            "type": "string"
          },
          "rating": {
            "example": 4,
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          },
          "title": {
            "example": "Updated title",
            "maxLength": 120,
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Reviews, verified purchases and the rating rollup",
    "title": "Reviews Service",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "AppController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/ready": {
      "get": {
        "operationId": "AppController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/reviews/admin/republish": {
      "post": {
        "operationId": "ModerationController_republish",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Re-emit product.rating_changed for all ratings (staff replay)",
        "tags": [
          "moderation"
        ]
      }
    },
    "/api/v1/reviews/me": {
      "get": {
        "operationId": "ReviewsController_listMine",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "List my authored reviews",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/moderation": {
      "get": {
        "operationId": "ModerationController_listModeration",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter reviews by moderation status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedReviewsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "List reviews in moderation queue (staff)",
        "tags": [
          "moderation"
        ]
      }
    },
    "/api/v1/reviews/products/{productId}": {
      "get": {
        "operationId": "ReviewsController_listForProduct",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedReviewsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "List approved reviews for a product (public, newest first)",
        "tags": [
          "reviews"
        ]
      },
      "post": {
        "operationId": "ReviewsController_create",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Submit a review for a purchased product",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/products/{productId}/eligibility": {
      "get": {
        "operationId": "ReviewsController_eligibility",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibilityResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Check if current customer can review this product",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{id}": {
      "patch": {
        "operationId": "ReviewsController_update",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Edit an existing review (owner only; resets status to pending)",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{id}/approve": {
      "post": {
        "operationId": "ModerationController_approve",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Approve a review (staff)",
        "tags": [
          "moderation"
        ]
      }
    },
    "/api/v1/reviews/{id}/reject": {
      "post": {
        "operationId": "ModerationController_reject",
        "parameters": [
          {
            "description": "Set by the gateway from the verified JWT.",
            "in": "header",
            "name": "x-user-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Reject a review (staff)",
        "tags": [
          "moderation"
        ]
      }
    }
  },
  "servers": [],
  "tags": []
}